+++ pyecharts/base.py
@@ -1,14 +1,17 @@
 # coding=utf-8
-
+import os
 import uuid
 import warnings
 
 from jinja2 import Markup
 
-import pyecharts.constants as constants
-import pyecharts.engine as engine
 import pyecharts.utils as utils
+import pyecharts.engine as engine
+import pyecharts.constants as constants
+import pyecharts.exceptions as exceptions
 from pyecharts.conf import CURRENT_CONFIG
+from pyecharts_javascripthon.api import TRANSLATOR
+from pyecharts.echarts.option import get_all_options
 
 
 class Base(object):
@@ -16,11 +19,13 @@ class Base(object):
     `Base`类是所有图形类的基类，提供部分初始化参数和基本的方法
     """
 
-    def __init__(self,
-                 width=800,
-                 height=400,
-                 renderer=constants.CANVAS_RENDERER,
-                 page_title=constants.PAGE_TITLE):
+    def __init__(
+        self,
+        width=800,
+        height=400,
+        renderer=constants.CANVAS_RENDERER,
+        page_title=constants.PAGE_TITLE,
+    ):
         """
 
         :param width:
@@ -40,6 +45,7 @@ class Base(object):
         self.renderer = renderer
         self._page_title = page_title
         self._js_dependencies = {'echarts'}
+        self.event_handlers = {}
 
     @property
     def chart_id(self):
@@ -61,10 +67,14 @@ class Base(object):
     def page_title(self):
         return self._page_title
 
+    def on(self, event_name, handler):
+        self.event_handlers[event_name] = handler
+
     def print_echarts_options(self):
         """ 打印输出图形所有配置项
         """
-        print(utils.json_dumps(self._option, indent=4))
+        snippet = TRANSLATOR.translate(self.options)
+        print(snippet.as_snippet())
 
     def show_config(self):
         """ 打印输出图形所有配置项
@@ -72,7 +82,7 @@ class Base(object):
         deprecated_tpl = 'The {} is deprecated, please use {} instead!'
         warnings.warn(
             deprecated_tpl.format('show_config', 'print_echarts_options'),
-            DeprecationWarning
+            DeprecationWarning,
         )
         self.print_echarts_options()
 
@@ -80,7 +90,7 @@ class Base(object):
         """ 渲染图表的所有配置项，为 web pages 服务，不过需先提供
         所需要的js 依赖文件
         """
-        env = engine.create_default_environment()
+        env = engine.create_default_environment(constants.DEFAULT_HTML)
         html = env.render_container_and_echarts_code(self)
         return Markup(html)
 
@@ -89,18 +99,22 @@ class Base(object):
         """
         return CURRENT_CONFIG.produce_html_script_list(self._js_dependencies)
 
-    def render(self,
-               path='render.html',
-               template_name='simple_chart.html',
-               object_name='chart',
-               extra_context=None):
-        env = engine.create_default_environment()
+    def render(
+        self,
+        path='render.html',
+        template_name='simple_chart.html',
+        object_name='chart',
+        **kwargs
+    ):
+        _, ext = os.path.splitext(path)
+        _file_type = ext[1:]
+        env = engine.create_default_environment(_file_type)
         env.render_chart_to_file(
             chart=self,
             object_name=object_name,
             path=path,
             template_name=template_name,
-            extra_context=extra_context
+            **kwargs
         )
 
     @staticmethod
@@ -138,23 +152,78 @@ class Base(object):
                 v_lst.append(seq[key])
         return k_lst, v_lst
 
+    def render_notebook(self):
+        warnings.warn(
+            'Implementation has been removed. ' +
+            'Please pass the chart instance directly to Jupyter.' +
+            'If you need more help, please read documentation'
+        )
+
+    def _get_all_options(self, **kwargs):
+        return get_all_options(**kwargs)
+
     def _repr_html_(self):
         """ 渲染配置项并将图形显示在 notebook 中
         chart/page => charts
         chart.js_dependencies => require_config => config_items, libraries
         :return A unicode string.
         """
+        if CURRENT_CONFIG.jupyter_presentation != constants.DEFAULT_HTML:
+            return None
+
         require_config = CURRENT_CONFIG.produce_require_configuration(
             self.js_dependencies
         )
         config_items = require_config['config_items']
         libraries = require_config['libraries']
-        env = engine.create_default_environment()
+        env = engine.create_default_environment(constants.DEFAULT_HTML)
         return env.render_chart_to_notebook(
-            charts=(self,),
-            config_items=config_items,
-            libraries=libraries
+            charts=(self,), config_items=config_items, libraries=libraries
         )
+
+    def _repr_svg_(self):
+        content = self._render_as_image(constants.SVG)
+        if content:
+            # fix alignment problem in notebook
+            content = content.replace('position: absolute;', '')
+        return content
+
+    def _repr_png_(self):
+        return self._render_as_image(constants.PNG)
+
+    def _repr_jpeg_(self):
+        return self._render_as_image(constants.JPEG)
+
+    def _render_as_image(self, file_type):
+        """
+        This is an internal function to serve _repr_jpeg_,
+        _repr_png_ and _repr_svg_.
+
+        :param file_type: the parameter is mostly image file types.
+        """
+        if CURRENT_CONFIG.jupyter_presentation != file_type:
+            return None
+
+        if self.renderer == constants.SVG_RENDERER:
+            if file_type != constants.SVG:
+                raise exceptions.InvalidConfiguration(
+                    "svg renderer produces only svg image."
+                )
+
+        elif file_type not in [constants.JPEG, constants.PNG]:
+            # CANVAS_RENDERER here
+            raise exceptions.InvalidConfiguration(
+                "svg output requires svg renderer."
+            )
+
+        env = engine.create_default_environment(file_type)
+        outfile = 'tmp.' + file_type
+        content = env.render_chart_to_file(
+            chart=self, path=outfile, verbose=False
+        )
+        if content:
+            os.unlink(outfile)
+        return content
 
     def _add_chinese_map(self, map_name_in_chinese):
         name_in_pinyin = CURRENT_CONFIG.chinese_to_pinyin(map_name_in_chinese)