+++ pyecharts/charts/geo.py
@@ -1,10 +1,7 @@
 # coding=utf-8
 
-import copy
-
 from pyecharts.chart import Chart
-from pyecharts.option import get_all_options
-from pyecharts.constants import CITY_GEO_COORDS
+from pyecharts.datasets.coordinates import get_coordinate
 
 
 class Geo(Chart):
@@ -13,22 +10,55 @@ class Geo(Chart):
 
     地理坐标系组件用于地图的绘制，支持在地理坐标系上绘制散点图，线集。
     """
+
     def __init__(self, title="", subtitle="", **kwargs):
         super(Geo, self).__init__(title, subtitle, **kwargs)
+        self._coordinates = {}
+
+    def add_coordinate(self, name, longitude, latitude):
+        """
+        Add a geo coordinate for a position.
+        :param name: The name of a position
+        :param longitude: The longitude of coordinate.
+        :param latitude: The latitude of coordinate.
+        :return:
+        """
+        self._coordinates.update({name: [longitude, latitude]})
+
+    def get_coordinate(self, name, raise_exception=False):
+        """
+        Return coordinate for the city name.
+        :param name: City name or any custom name string.
+        :param raise_exception: Whether to raise exception if not exist.
+        :return: A list like [longitude, latitude] or None
+        """
+        if name in self._coordinates:
+            return self._coordinates[name]
+
+        coordinate = get_coordinate(name)
+        if coordinate is None and raise_exception:
+            raise ValueError('No coordinate is specified for {}'.format(name))
+
+        return coordinate
 
     def add(self, *args, **kwargs):
         self.__add(*args, **kwargs)
 
-    def __add(self, name, attr, value,
-              type="scatter",
-              maptype='china',
-              symbol_size=12,
-              border_color="#111",
-              geo_normal_color="#323c48",
-              geo_emphasis_color="#2a333d",
-              geo_cities_coords=None,
-              is_roam=True,
-              **kwargs):
+    def __add(
+        self,
+        name,
+        attr,
+        value,
+        type="scatter",
+        maptype='china',
+        symbol_size=12,
+        border_color="#111",
+        geo_normal_color="#323c48",
+        geo_emphasis_color="#2a333d",
+        geo_cities_coords=None,
+        is_roam=True,
+        **kwargs
+    ):
         """
 
         :param name:
@@ -66,74 +96,72 @@ class Geo(Chart):
         """
         assert len(attr) == len(value)
         kwargs.update(type="geo")
-        chart = get_all_options(**kwargs)
+        chart = self._get_all_options(**kwargs)
 
         if geo_cities_coords:
-            _geo_cities_coords = copy.deepcopy(geo_cities_coords)
-        else:
-            _geo_cities_coords = copy.deepcopy(CITY_GEO_COORDS)
+            for name, coord in geo_cities_coords.items():
+                self.add_coordinate(name, coord[0], coord[1])
 
         _data = []
         for _name, _value in zip(attr, value):
-            if _name in _geo_cities_coords:
-                city_coordinate = _geo_cities_coords.get(_name)
-                city_coordinate.append(_value)
-                _data.append({"name": _name, "value": city_coordinate})
-            else:
-                print("%s coordinates is not found" % _name)
+            _coordinate = self.get_coordinate(_name, raise_exception=True)
+            _data_value = [_coordinate[0], _coordinate[1], _value]
+            _data.append({"name": _name, "value": _data_value})
         self._option.update(
             geo={
                 "map": maptype,
                 "roam": is_roam,
                 "label": {
-                    "emphasis": {
-                        "show": True,
-                        "textStyle": {
-                            "color": "#eee"
-                        }
-                    }},
+                    "emphasis": {"show": True, "textStyle": {"color": "#eee"}}
+                },
                 "itemStyle": {
                     "normal": {
                         "areaColor": geo_normal_color,
-                        "borderColor": border_color
+                        "borderColor": border_color,
                     },
-                    "emphasis": {
-                        "areaColor": geo_emphasis_color
-                    }}
-            })
+                    "emphasis": {"areaColor": geo_emphasis_color},
+                },
+            }
+        )
         self._option.get('legend')[0].get('data').append(name)
 
         if type == "scatter":
-            self._option.get('series').append({
-                "type": type,
-                "name": name,
-                "coordinateSystem": 'geo',
-                "symbol": chart['symbol'],
-                "symbolSize": symbol_size,
-                "data": _data,
-                "label": chart['label'],
-            })
+            self._option.get('series').append(
+                {
+                    "type": type,
+                    "name": name,
+                    "coordinateSystem": 'geo',
+                    "symbol": chart['symbol'],
+                    "symbolSize": symbol_size,
+                    "data": _data,
+                    "label": chart['label'],
+                }
+            )
 
         elif type == "effectScatter":
-            self._option.get('series').append({
-                "type": type,
-                "name": name,
-                "coordinateSystem": 'geo',
-                "showEffectOn": "render",
-                "rippleEffect": chart['effect'],
-                "symbol": chart['symbol'],
-                "symbolSize": symbol_size,
-                "data": _data,
-                "label": chart['label'],
-            })
+            self._option.get('series').append(
+                {
+                    "type": type,
+                    "name": name,
+                    "coordinateSystem": 'geo',
+                    "showEffectOn": "render",
+                    "rippleEffect": chart['effect'],
+                    "symbol": chart['symbol'],
+                    "symbolSize": symbol_size,
+                    "data": _data,
+                    "label": chart['label'],
+                }
+            )
 
         elif type == "heatmap":
-            self._option.get('series').append({
-                "type": type,
-                "name": name,
-                "coordinateSystem": 'geo',
-                "data": _data,
-            })
+            self._option.get('series').append(
+                {
+                    "type": type,
+                    "name": name,
+                    "coordinateSystem": 'geo',
+                    "data": _data,
+                }
+            )
 
         self._add_chinese_map(maptype)
         self._config_components(**kwargs)