+++ pyecharts/charts/scatter3D.py
@@ -1,7 +1,6 @@
 # coding=utf-8
 
 from pyecharts.chart import Chart
-from pyecharts.option import get_all_options
 import pyecharts.constants as constants
 
 
@@ -18,9 +17,7 @@ class Scatter3D(Chart):
     def add(self, *args, **kwargs):
         self.__add(*args, **kwargs)
 
-    def __add(self, name, data,
-              grid3d_opacity=1,
-              **kwargs):
+    def __add(self, name, data, grid3d_opacity=1, **kwargs):
         """
 
         :param name:
@@ -31,26 +28,26 @@ class Scatter3D(Chart):
             3D 笛卡尔坐标系组的透明度（点的透明度），默认为 1，完全不透明。
         :param kwargs:
         """
-        kwargs.update(xaxis3d_type='value',
-                      yaxis3d_type='value',
-                      zaxis3d_type='value')
-        chart = get_all_options(**kwargs)
+        kwargs.update(
+            xaxis3d_type='value', yaxis3d_type='value', zaxis3d_type='value'
+        )
+        chart = self._get_all_options(**kwargs)
 
         self._option.get('legend')[0].get('data').append(name)
         self._option.update(
             xAxis3D=chart['xaxis3D'],
             yAxis3D=chart['yaxis3D'],
             zAxis3D=chart['zaxis3D'],
-            grid3D=chart['grid3D']
+            grid3D=chart['grid3D'],
         )
 
-        self._option.get('series').append({
-            "type": "scatter3D",
-            "name": name,
-            "data": data,
-            "label": chart['label'],
-            "itemStyle": {
-                "opacity": grid3d_opacity
+        self._option.get('series').append(
+            {
+                "type": "scatter3D",
+                "name": name,
+                "data": data,
+                "label": chart['label'],
+                "itemStyle": {"opacity": grid3d_opacity},
             }
-        })
+        )
         self._config_components(**kwargs)